requirements.txt
markdown2
gunicorn
django-heroku